solved conflict issues
