+++ src/app/universities/[id]/Certcard.jsx
@@ -1,15 +1,19 @@
 import React from 'react'
 import Image from "next/image";
 
-const Certcard = ({ imgSrc }) => {
+const Certcard = ({ imgSrc, title }) => {
     return (
-      <div className="max-w-xs bg-white flex flex-col items-center text-center">
+      <div className="max-w-xs bg-white py-6 flex flex-col items-center text-center">
         {/* Image */}
-        <div className="lg:w-[250px] lg:h-[120px] mb-4">
-          <Image src={imgSrc} alt={imgSrc} className="w-full h-full object-contain" />
+        <div className="w-24 h-24 mb-4">
+          <Image src={imgSrc} width={400} height={400} alt={title} className="w-full h-full object-contain" />
         </div>
+  
+        {/* Text */}
+        <h3 className="text-[14px] md:text-[16px] lg:text-[20px] font-semibold text-gray-800">{title}</h3>
+       
       </div>
     );
   };
 
-export default Certcard
+export default Certcard
